If there are problems with running pdb4amber, a separate input file is written which upon execution will provide files required for next steps.

+++ mdms/topology_prep.py
@@ -244,7 +244,7 @@ def ligands_parameters():
             save_to_file(f"ligands_multiplicities = {lig_multiplicities}\n", filename)
 
 
-def antechamber_parmchk_input():
+def                antechamber_parmchk_input():
     # finding ligands residues in control file
     control = read_file(filename)
     ligands = r'ligands\s*=\s*\[(.*)\]'
@@ -300,30 +300,59 @@ def antechamber_parmchk_input():
         else:
             # try to find ligands_processing entry in the control file
             # try to find if there is a ligands_pdb4amber_inputs entry in the control file
-            ligands = r'ligands_pdb4amber_inputs\s*=\s*\[(.*)\]'
-            ligands_match = re.search(ligands, control)
-            if ligands_match:
-
-
-            print(f"\n!!WARNING!!\n"
-                  f"There were some problems with running pdb4amber from within MDMS.\n"
-                  f"If you installed Ambertools previously and it suddenly stopped working, it might be due to "
-                  f"the clash of Python versions - MDMS uses Python 3.6, whereas pdb4amber was written in Python 2.7.\n"
-                  f"This issue usually occurs on HPC facilities which uses environmental modules feature alongside "
-                  f"your own installation of Python.\n"
-                  f"If that might be true in your case, for a moment please choose Python 2.7 as a default Python"
-                  f"interpreter (i. e. by loading the appropriate module).\n"
-                  f"Please also make sure that pdb4amber is installed correctly (just follow Amber Manual on how to "
-                  f"install Ambertools).\nAt this point you should be able to use pdb4amber in the terminal.\n"
-                  f"In such case, just execute the following "
-                  f"In such case, just copy the content of pdb4amber.in into "
-                  f"terminal and press enter).\n")
-            pdb_process_input = Path('pdb4amber.in')
-            pdb_process_input_path = Path(pdb_process_input)
-            if pdb_process_input_path.exists():
-                os.remove(pdb_process_input_path)
-            with open(pdb_process_input, 'w') as file:
-                file.write(pdb4amber_input)
+            ligands_inputs = r'ligands_pdb4amber_inputs\s*=\s*\[(.*)\]'
+            ligands_inputs_match = re.search(ligands, control)
+            if ligands_inputs_match:
+                # search for outputs from pdb4amber - they must have been processed manually; if they exist, everthing
+                # is fine; if they don't print info on how to proceed again; every ligand in the list is checked
+                for x in range(0, len(ligands_list)):
+                    lig_path = Path(f'{ligands_list[x]}.pdb')
+                    if lig_path.exists():
+                        continue
+                    else:
+                        print(f"\nYou have not processed ligands' files with pdb4amber yet\n"
+                              f"There were some problems with running pdb4amber from within MDMS.\n"
+                              f"If you installed Ambertools previously and it suddenly stopped working, it might be due to "
+                              f"the clash of Python versions - MDMS uses Python 3.6, whereas pdb4amber was written in Python 2.7.\n"
+                              f"This issue usually occurs on HPC facilities which uses environmental modules feature alongside "
+                              f"your own installation of Python.\n"
+                              f"If that might be true in your case, for a moment please choose Python 2.7 as a default Python"
+                              f"interpreter (i. e. by loading the appropriate module).\n"
+                              f"Please also make sure that pdb4amber is installed correctly (just follow Amber Manual on how to "
+                              f"install Ambertools).\nAt this point you should be able to use pdb4amber in the terminal.\n"
+                              f"In such case, just execute the following "
+                              f"In such case, just copy the content of pdb4amber.in into "
+                              f"terminal and press enter).\n")
+                        # just single printing of the prompt should be enough, therefore it is halted here
+                        break
+            else:
+                print(f"\n!!WARNING!!\n"
+                      f"There were some problems with running pdb4amber from within MDMS.\n"
+                      f"If you installed Ambertools previously and it suddenly stopped working, it might be due to "
+                      f"the clash of Python versions - MDMS uses Python 3.6, whereas pdb4amber was written in Python 2.7.\n"
+                      f"This issue usually occurs on HPC facilities which uses environmental modules feature alongside "
+                      f"your own installation of Python.\n"
+                      f"If that might be true in your case, for a moment please choose Python 2.7 as a default Python"
+                      f"interpreter (i. e. by loading the appropriate module).\n"
+                      f"Please also make sure that pdb4amber is installed correctly (just follow Amber Manual on how to "
+                      f"install Ambertools).\nAt this point you should be able to use pdb4amber in the terminal.\n"
+                      f"In such case, just copy the content of each line of pdb4amber.in file into terminal and press"
+                      f"enter. As a result, you will obtain processed PDB ligands' file which will be ready for further"
+                      f"steps.")
+                pdb_process_input = ('pdb4amber.in')
+                pdb_process_input_path = Path(pdb_process_input)
+                if pdb_process_input_path.exists():
+                    os.remove(pdb_process_input_path)
+                for x in range(0, len(ligands_list)):
+                    # copying original ligand PDB file - output from pdb4amber will be
+                    # supplied to antechamber and parmchk
+                    ligand_copy = f"cp {ligands_list[x]}_raw.pdb {ligands_list[x]}_original.pdb"
+                    subprocess.run([f"{ligand_copy}"], shell=True)
+                    # input for pdb4amber
+                    pdb4amber_input = f"pdb4amber -i {ligands_list[x]}_original.pdb -o {ligands_list[x]}.pdb "
+                    # input for pdb4amber appended to the file
+                    with open(pdb_process_input, 'a') as file:
+                        file.write(pdb4amber_input)
         # creating antechamber and parmchk inputs
         for x in range(0, len(ligands_list)):
             # input for antechamber
